Refactor code to handle Date and simplify Strings
The major modification was changing the data type of dates in the Game and User models from String to the Date type. This leads to a more straightforward and object-oriented way of handling dates in the code. All database queries and data inserts have been adjusted accordingly as well. Additionally, long String variables storing SQL queries were simplified by directly including the statements in the method's code, enhancing readability and maintainability of the code.